// web/src/api/generated/models/updateUserRequest.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * User Management API
 * OpenAPI spec version: 0.0.0
 */

/**
 * Update user request
 */
export interface UpdateUserRequest {
  /**
   * User name
   * @minLength 1
   * @maxLength 100
   */
  name?: string;
  /** User email address */
  email?: string;
}